+++ CHANGELOG.md
@@ -4,6 +4,63 @@ A breaking change will get clearly marked in this log.
 
 ## Unreleased
 
+## [v5.0.0](https://github.com/stellar/js-stellar-sdk/compare/v4.1.0...v5.0.0)
+
+### Add
+- Add fee bump related attributes to `TransactionResponse` ([#532](https://github.com/stellar/js-stellar-sdk/pull/532)): 
+    - `fee_account: string`.
+    - `fee_bump_transaction: FeeBumpTransactionResponse`:
+      ```js
+      interface FeeBumpTransactionResponse {
+        hash: string;
+        signatures: string[];
+      }
+      ```
+    - `inner_transaction: InnerTransactionResponse`: 
+      ```js
+      interface InnerTransactionResponse {
+        hash: string;
+        signatures: string[];
+        max_fee: string;
+      }
+      ```
+- Add `memo_bytes: string` to `TransactionResponse` ([#532](https://github.com/stellar/js-stellar-sdk/pull/532)).
+- Add `authorize_to_maintain_liabilities: boolean` to `AllowTrustOperation` ([#532](https://github.com/stellar/js-stellar-sdk/pull/532)).
+- Add `is_authorized_to_maintain_liabilities: boolean` to `BalanceLineNative` ([#532](https://github.com/stellar/js-stellar-sdk/pull/532)).
+- Add new result codes to `TransactionFailedResultCodes` ([#531](https://github.com/stellar/js-stellar-sdk/pull/531)).
+  ```js
+  TX_FEE_BUMP_INNER_SUCCESS = "tx_fee_bump_inner_success",
+  TX_FEE_BUMP_INNER_FAILED = "tx_fee_bump_inner_failed",
+  TX_NOT_SUPPORTED = "tx_not_supported",
+  TX_SUCCESS = "tx_success",
+  TX_TOO_EARLY = "tx_too_early",
+  TX_TOO_LATE = "tx_too_late",
+  TX_MISSING_OPERATION = "tx_missing_operation",
+  TX_INSUFFICIENT_BALANCE = "tx_insufficient_balance",
+  TX_NO_SOURCE_ACCOUNT = "tx_no_source_account",
+  TX_INSUFFICIENT_FEE = "tx_insufficient_fee",
+  TX_INTERNAL_ERROR = "tx_internal_error",
+  ```
+
+### Breaking changes 
+- The attributes `max_fee` and `fee_charged` in `TransactionResponse` can be now a `number` or a `string`. 
+  Update your code to handle both types since Horizon will start sending `string` in version `1.3.0` ([#528](https://github.com/stellar/js-stellar-sdk/pull/528)).
+- Bump `stellar-base` to `v3.0.0`: This new version of stellar-base brings support for protocol 13, including multiple breaking changes which might affect your code, please review the list of breaking changes in [stellar-base@3.0.0](https://github.com/stellar/js-stellar-base/releases/tag/v3.0.0) release ([#524](https://github.com/stellar/js-stellar-sdk/pull/524)).
+- Make `networkPassphrase` a required argument in `Utils.buildChallengeTx` and  `Utils.readChallengeTx` ([#524](https://github.com/stellar/js-stellar-sdk/pull/524)).
+- Remove `Server.paths` ([#525](https://github.com/stellar/js-stellar-sdk/pull/525)).
+
+## [v5.0.0-alpha.2](https://github.com/stellar/js-stellar-sdk/compare/v5.0.0-alpha.1..v5.0.0-alpha.2)
+
+### Update
+- Update `stellar-base` to `v3.0.0-alpha-1`.
+
+## [v5.0.0-alpha.1](https://github.com/stellar/js-stellar-sdk/compare/v4.1.0...v5.0.0-alpha.1)
+
+### Breaking changes
+- Bump `stellar-base` to `v3.0.0-alpha-0`: This new version of stellar-base brings support for protocol 13, including multiple breaking changes which might affect your code, please review the list of breaking changes in [stellar-base@3.0.0-alpha.0](https://github.com/stellar/js-stellar-base/releases/tag/v3.0.0-alpha.0) release ([#524](https://github.com/stellar/js-stellar-sdk/pull/524)).
+- Make `networkPassphrase` a required argument in `Utils.buildChallengeTx` and  `Utils.readChallengeTx` ([#524](https://github.com/stellar/js-stellar-sdk/pull/524)).
+- Remove `Server.paths` ([#525](https://github.com/stellar/js-stellar-sdk/pull/525)).
+
 ## [v4.1.0](https://github.com/stellar/js-stellar-sdk/compare/v4.0.2...v4.1.0)
 
 ### Add
@@ -11,7 +68,7 @@ A breaking change will get clearly marked in this log.
 
   Extends `server.submitTransaction` to always run a memo required check before
   sending the transaction.  If any of the destinations require a memo and the
-  transaction doesn't include one, then an `AccountRequiresMemoError` will be thrown.  
+  transaction doesn't include one, then an `AccountRequiresMemoError` will be thrown.
 
   You can skip this check by passing `{skipMemoRequiredCheck: true}` to `server.submitTransaction`: